+++ packages/adhere-ui-reactquill-sandbox/lib/index.d.ts
@@ -1,0 +1,4 @@
+import '@webcomponents/webcomponentsjs/custom-elements-es5-adapter.js';
+import '@webcomponents/webcomponentsjs/webcomponents-loader.js';
+import ReactQuillSandbox from './reactQuillSandbox';
+export default ReactQuillSandbox;