+++ packages/adhere-util-resource/es/dict/dict.flow.config.js
@@ -1,2 +1,0 @@
-export default{initStatic:function(){},initRemote:function(){}};
-//# sourceMappingURL=dict.flow.config.js.map